requirements.txt
aws_cdk.aws_dynamodb
aws_cdk.core
boto3
cdk-chalice
chalice
